Create store for retrieving list of collections
Abstracted getting the list of collections from a user into a Store
(DumbWatchlistOnlyCollectionsListStore). Right now is a really dumb one. I've
put all the common behavior into a common store (parallel to what we did with
CollectionStore) so that we can substitute easily
DumbWatchlistOnlyCollectionsListStore with the real storage class once we do
it.

As a consequence of cleaning up the CollectionsList model, it turned out we
don't actually store anything extra on that model, so it can actually be
removed and just use an array of Collection for representing a list of
collections, so that's simpler.

Change-Id: Icdac5fc5ab39a696a0c0a142f1f97e094da74029