Split the Collection model and add JSON serialization
And also add JSON serialization on the models.

On the list of collections page it will use CollectionInfo which is the base
model plus a count attribute.

On the details page of a collection it will use the full Collection, which has
the base plus the items of the collection fully expanded.

This should work but is part of a bigger patch for creating the storage for the
index file. Check previous and next patches also.

Change-Id: Ia36e9d61d771241f55aa8202f7a9dcf6f8dd3f64