Hygiene: Collection stores as functions & fix docs
In preparation for more API work:

I've refactored Collection, UserPageCollection and WatchlistCollection to be
static functions that take parameters and return results. Less object oriented
stuff since oo makes the design for adding, updating and creating hard from
what we where doing.

Also, we should use for collection loading/saving a implementation of
stores\CollectionStorage, right now the only one is UserPageCollection but Rob
is working on DBCollection. This classes should handle internally the
difference between watchlist loading on its hardcoded id (see
UserPageCollection::getByUserAndId for an example).

Also fixed a few of small fixes to php docs that i've noticed form using
phpstorm.

I'm going to refactor the stores\UserPageCollectionsList to follow the same
style as these ones too, for easier handling on the API and a clearer interface
(in following patches)

Change-Id: I09eb6ef45bbdeea9d5f4cf6b2647d8f97631e2fc