Hygiene: Clean up stores\UserPageCollectionsList
Cleaned up the stores for CollectionsListStorage and UserPageCollectionsList to
follow the Collection store model (previous patch). CollectionsList storage
stores will implement the interface CollectionsListStorage (right now only
UserPageCollectionsList but soon DBCollectionsList).

Introduced the CollectionsList model, which is only an array of CollectionInfo
but it encapsulates the array and the logic for adding private/public items.
(all of this was mangled in the collectionslist store before).

Refactored properly everywhere that those were used.

Cleaned up a couple of things I've seen from php warnings.

Change-Id: Iada0d5d84f1f52026c5a7f7a5ce68f2380d000f8