Update gems and fix mediawiki-selenium version
Needed a `bundle update` for updating the selenium and watir drivers (were
broken on phantom and latest firefox) but I needed to fix mediawiki-selenium to
0.4.2 since 1.0 is published but incompatible so updating without fixing that
version would break running the browser tests.

We'll need to update to mediawiki-selenium >=1 at some point following the
upgrade instructions at some point in the near future.

Bug: T93195
Change-Id: I765c88c431d35b816ff24456aae441c7f260ff51