Design: Improve spacing of content overlay on phones
Reduced spacing by making it percent based on phones instead of fixed em.
Result is better spacing on phones based on the screen size and even better
when phone keyboard is up with the overlay open (see bug for bad behavior).

See bug for screenshots of how this patch looks like.

Bug: T95628
Change-Id: I50bf8200fc462d4508f79400833c6bb932b65e27