When creating a collection show CTA && toasts after edited
When creating a new collection from an article page, events were emitted before
redirecting to the editor. Because of that, when you finish with the editor no
toast would be seen.

Now when you add an article to a collection, if you create a new one then
you'll go to the editor, and when finished you'll see the toast and CTA
triggered.

Bug: T98316
Change-Id: Idfdd38dd799f498d4d36cc9e92b247ef36cae1d5