Extend the core TemplateParser
TemplateParser#getTemplate eval's a compiled template so that the
renderer function can be returned to the caller. Due to a bug in HHVM's
implementation of eval [0], we can't rely on the render function
returned by #getTemplate on being the correct one.

Extend the Gather\views\TemplateParser class so that #getTemplate will
write the compiled template to a file that is then required, thus
avoiding eval. The compiled templates are written to the
templates/compiled directory and should be committed whenever they
change.

Since the new TemplateParser class can't change the public API of the
parent TemplateParser (LSP), the rest of the codebase should be agnostic
to the change. However, because we're expecting the compiled templates
to be committed, modify Gather\views\helpers\Template::render
to force template recompilation whenever the GatherRecompileTemplates
config variable is true.

The GatherRecompileTemplates config variable defaults to false.
Developers are expected to override it during development.

Also reverts "Work around Mustache problems" (1c1f658) so we're in a
clean state immediately.

[0] https://github.com/facebook/hhvm/pull/5502

Bug: T102941
Change-Id: I729dd34d102879e596428cb07d8cf381071068bb