Fix collections list pagination
Infinite scrolling is broken since the template has changed for use on the
server but not on javascript.

This patch takes into consideration that a CollectionsList may want to show an
owner in the items if the collection doesn't actually have one (as we've done
on the server).

Also the collection mode is set on the server, and then rendered on the
template so that the client can pick up which mode the collection has and make
the pagination api calls appropiately.

How to test:

* Change the limit to 5, ensure you have a user with more than 5 collections and
that in your instance you have more than 5 collections "active" (more than
4 items).

* Visit /Special:Gather/by/YOURUSER. Let it load. Scroll to the bottom, see the
new collections with the proper privacy status under the title.

* Visit /Special:Gather/all/active. Let it load. Scroll to the bottom, see the
new collections with the owner info under the title instead of the privacy
status.

Bug: T102983
Change-Id: Idd0b21a55455192037f68b517e2ead00d35a70cf